Changed Activities to Fragments
Implemented fragments and navigation with a NavController for future updates
