+++ .idea/misc.xml
@@ -5,6 +5,8 @@
       <map>
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/activity_add_medicament.xml" value="0.35104166666666664" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/activity_main.xml" value="0.35104166666666664" />
+        <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/fragment_add.xml" value="0.1" />
+        <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/fragment_list.xml" value="0.33" />
         <entry key="..\:/Users/Ardak/Documents/GitHub/healthnote-android/app/src/main/res/layout/text_row_item.xml" value="0.33" />
         <entry key="..\:/Users/z0045evs/Desktop/AndroidDev/healthnote-android/app/src/main/res/layout/activity_add_medicament.xml" value="0.25" />
         <entry key="..\:/Users/z0045evs/Desktop/AndroidDev/healthnote-android/app/src/main/res/layout/activity_main.xml" value="0.2" />
