+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.HealthNote">
         <activity
-            android:name=".AddMedicamentActivity"
-            android:exported="false"
-            android:parentActivityName=".MainActivity"/>
-        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
