+++ app/src/main/java/de/othregensburg/healthnote/fragments/add/AddFragment.kt
@@ -1,28 +1,35 @@
 package de.othregensburg.healthnote.fragments.add
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
+import android.text.TextUtils
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
+import android.widget.Toast
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.fragment.findNavController
 import de.othregensburg.healthnote.R
+import de.othregensburg.healthnote.data.Medicament
 import de.othregensburg.healthnote.data.MedicamentViewModel
+import de.othregensburg.healthnote.databinding.FragmentAddBinding
 
 class AddFragment : Fragment() {
 
     private lateinit var mMedViewModel: MedicamentViewModel
+    private var _binding: FragmentAddBinding? = null
+    private val binding get() = _binding!!
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
+        _binding = FragmentAddBinding.inflate(inflater, container, false)
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_add, container, false)
+        val view = binding.root
 
-        mMedViewModel = ViewModelProvider(this).get(MedicamentViewModel::class.java)
+        mMedViewModel = ViewModelProvider(this)[MedicamentViewModel::class.java]
 
-        view.findViewById<Button>(R.id.add_button).setOnClickListener {
+        binding.checkMarkButton.setOnClickListener {
             insertDataToDatabase()
         }
 
@@ -30,6 +37,18 @@ class AddFragment : Fragment() {
     }
 
     private fun insertDataToDatabase() {
-        val name = na
+        val name = binding.addMedicineName.text.toString()
+        if (inputCheck(name)) {
+            val med = Medicament(0, name)
+            mMedViewModel.addMed(med)
+            Toast.makeText(requireContext(), "Successfully added", Toast.LENGTH_SHORT).show()
+            findNavController().navigate(R.id.action_addFragment_to_listFragment)
+        } else {
+            Toast.makeText(requireContext(), "Please fill out all fields", Toast.LENGTH_SHORT).show()
+        }
+    }
+
+    private fun inputCheck(name: String) : Boolean {
+        return !(TextUtils.isEmpty(name))
     }
 }
